requirements.txt
frozen-flask
flask_flatpages
